auth-service/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
pyjwt==2.9.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
